requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
hpack==4.0.0
itsdangerous==2.2.0
keras==3.6.0
labelImg==1.8.6
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
munkres==1.1.4
mysql-connector-python==9.1.0
namex==0.0.8
networkx==2.8.4  
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
pandas==1.5.0
protobuf==4.25.5
Pygments==2.18.0
PyMySQL==1.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
rich==13.9.2
shiboken6==6.8.0.1
smmap==5.0.1
SQLAlchemy==2.0.36
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
ultralytics==8.3.19
ultralytics-thop==2.0.9
Werkzeug==3.0.4
wrapt==1.16.0
yolov4==3.2.0
zstandard==0.23.0
